Revert "Move deposit into strategy for operator to avs level"
This reverts commit 8a0571265857d05fe55e674d23382cd3608f7687.
