+++ examples/awesome-vault-service/contracts/script/deployments/core/31337.json
@@ -1,1 +1,1 @@
-{"lastUpdate":{"timestamp":"1747257881","block_number":"0"},"addresses":{"proxyAdmin":"0x5fbdb2315678afecb367f032d93f642f64180aa3","delegation":"0x9fe46736679d2d9a65f0992f2272de9f3c7fa6e0","delegationManagerImpl":"0x4ed7c70f96b99c776995fb64377f0d4ab3b0e1c1","avsDirectory":"0x610178da211fef7d417bc0e6fed39f05609ad788","avsDirectoryImpl":"0xa85233c63b9ee964add6f2cffe00fd84eb32338f","strategyManager":"0x0165878a594ca255338adfa4d48449f69242eb8f","strategyManagerImpl":"0x4a679253410272dd5232b3ff7cf5dbb88f295319","eigenPodManager":"0x3aa5ebb10dc797cac828524e59a333d0a371443c","eigenPodManagerImpl":"0xc5a5c42992decbae36851359345fe25997f5c42d","strategyFactory":"0x9a9f2ccfde556a7e9ff0848998aa4a0cfd8863ae","rewardsCoordinator":"0xa51c1fc2f0d1a1b8494ed1fe312d7c3a78ed91c0","pauserRegistry":"0x0b306bf915c4d645ff596e518faf3f9669b97016","strategyBeacon":"0x84ea74d481ee0a5332c457a4d796187f6ba67feb","allocationManager":"0x2279b7a0a67db372996a5fab50d91eaa73d2ebe6","permissionController":"0x59b670e9fa9d0a427751af201d676719a970857b"}}
+{"lastUpdate":{"timestamp":"1747247250","block_number":"0"},"addresses":{"proxyAdmin":"0x5fbdb2315678afecb367f032d93f642f64180aa3","delegation":"0x9fe46736679d2d9a65f0992f2272de9f3c7fa6e0","delegationManagerImpl":"0x4ed7c70f96b99c776995fb64377f0d4ab3b0e1c1","avsDirectory":"0x610178da211fef7d417bc0e6fed39f05609ad788","avsDirectoryImpl":"0xa85233c63b9ee964add6f2cffe00fd84eb32338f","strategyManager":"0x0165878a594ca255338adfa4d48449f69242eb8f","strategyManagerImpl":"0x4a679253410272dd5232b3ff7cf5dbb88f295319","eigenPodManager":"0x3aa5ebb10dc797cac828524e59a333d0a371443c","eigenPodManagerImpl":"0xc5a5c42992decbae36851359345fe25997f5c42d","strategyFactory":"0x9a9f2ccfde556a7e9ff0848998aa4a0cfd8863ae","rewardsCoordinator":"0xa51c1fc2f0d1a1b8494ed1fe312d7c3a78ed91c0","pauserRegistry":"0x0b306bf915c4d645ff596e518faf3f9669b97016","strategyBeacon":"0x84ea74d481ee0a5332c457a4d796187f6ba67feb","allocationManager":"0x2279b7a0a67db372996a5fab50d91eaa73d2ebe6","permissionController":"0x59b670e9fa9d0a427751af201d676719a970857b"}}
