+++ operator/registration.go
@@ -3,6 +3,7 @@ package operator
 import (
 	"context"
 	"errors"
+	"math/big"
 	"os"
 	"time"
 
@@ -10,6 +11,8 @@ import (
 	"github.com/Layr-Labs/eigensdk-go/crypto/bls"
 	"github.com/Layr-Labs/eigensdk-go/crypto/ecdsa"
 	"github.com/Layr-Labs/eigensdk-go/signerv2"
+
+	erc20mock "github.com/Layr-Labs/eigensdk-go/contracts/bindings/MockERC20"
 
 	"github.com/Layr-Labs/eigensdk-go/chainio/clients/elcontracts"
 	"github.com/Layr-Labs/eigensdk-go/chainio/clients/wallet"
@@ -25,6 +28,24 @@ import (
 )
 
 func RegisterOperatorOnStartup(c RegistrationConfig) error {
+	// operatorAddr := common.HexToAddress("0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266")
+	// allocationManagerAddr := common.HexToAddress("0x2279b7a0a67db372996a5fab50d91eaa73d2ebe6")
+	// avsAddress := common.HexToAddress("0xcd8a1c3ba11cf5ecfa6267617243239504a98d90")
+	// registryCoordinatorAddr := common.HexToAddress("0xfd471836031dc5108809d173a067e8486b9047a3")
+	// strategyAddr := common.HexToAddress("0x2b961e3959b79326a8e7f64ef0d2d825707669b5")
+	// ethHttpUrl := "http://localhost:8545"
+
+	// delegationManagerAddress := common.HexToAddress("0x9fe46736679d2d9a65f0992f2272de9f3c7fa6e0")
+	// rewardsCoordinatorAddress := common.HexToAddress("0xa51c1fc2f0d1a1b8494ed1fe312d7c3a78ed91c0")
+	// permissionControllerAddress := common.HexToAddress("0x59b670e9fa9d0a427751af201d676719a970857b")
+
+	// ecdsaKeyStorePath := "keys/test.ecdsa.key.json"
+	// blsKeyStorePath := "keys/test.bls.key.json"
+	// stringMintAmount := "1000000000000000000000"
+
+	// allocatableMagnitude := uint64(1000000000000000)
+	// operatorSetId := uint32(0)
+
 	ethRpcClient, err := ethclient.Dial(c.EthHttpUrl)
 	if err != nil {
 		c.Logger.Errorf("Cannot create http ethclient", "err", err)
@@ -81,6 +102,19 @@ func RegisterOperatorOnStartup(c RegistrationConfig) error {
 	)
 	if err != nil {
 		c.Logger.Fatalf("Failed to register operator with EigenLayer on startup: %v", err.Error())
+	}
+
+	err = DepositIntoStrategyForOperator(
+		c.Logger,
+		elcontractsConfig,
+		ethRpcClient,
+		c.StrategyAddr,
+		txMgr,
+		c.OperatorAddr,
+		c.AmountToMint,
+	)
+	if err != nil {
+		c.Logger.Fatalf("Failed to deposit into strategy for operator on startup: %v", err.Error())
 	}
 
 	blsKeyPassword, ok := os.LookupEnv("OPERATOR_BLS_KEY_PASSWORD")
@@ -237,6 +271,71 @@ func SetAllocationDelay(
 		"txHash",
 		receipt.TxHash.String(),
 	)
+
+	return nil
+}
+
+func DepositIntoStrategyForOperator(
+	logger logging.Logger,
+	elcontractsConfig elcontracts.Config,
+	ethClient *ethclient.Client,
+	strategyAddr common.Address,
+	txMgr txmgr.TxManager,
+	operatorAddr common.Address,
+	amount *big.Int,
+) error {
+	elReader, err := elcontracts.NewReaderFromConfig(elcontractsConfig, ethClient, logger)
+	if err != nil {
+		logger.Error("Error creating eigenlayer chain writer", "err", err)
+		return err
+	}
+
+	elWriter, err := elcontracts.NewWriterFromConfig(
+		elcontractsConfig,
+		ethClient,
+		logger,
+		&metrics.EigenMetrics{},
+		txMgr,
+	)
+	if err != nil {
+		logger.Error("Error creating eigenlayer chain writer", "err", err)
+		return err
+	}
+
+	_, tokenAddr, err := elReader.GetStrategyAndUnderlyingToken(context.Background(), strategyAddr)
+	if err != nil {
+		logger.Error("Failed to fetch strategy contract", "err", err)
+		return err
+	}
+	logger.Info(tokenAddr.String())
+
+	contractErc20Mock, err := erc20mock.NewContractMockERC20(tokenAddr, ethClient)
+	if err != nil {
+		logger.Error("Failed to fetch ERC20Mock contract", "err", err)
+		return err
+	}
+	txOpts, err := txMgr.GetNoSendTxOpts()
+	if err != nil {
+		logger.Errorf("Error in GetNoSendTxOpts")
+		return err
+	}
+
+	tx, err := contractErc20Mock.Mint(txOpts, operatorAddr, amount)
+	if err != nil {
+		logger.Errorf("Error assembling Mint tx")
+		return err
+	}
+	_, err = txMgr.Send(context.Background(), tx, true)
+	if err != nil {
+		logger.Errorf("Error submitting Mint tx")
+		return err
+	}
+
+	_, err = elWriter.DepositERC20IntoStrategy(context.Background(), strategyAddr, amount, true)
+	if err != nil {
+		logger.Errorf("Error depositing into strategy", "err", err)
+		return err
+	}
 
 	return nil
 }
